requirements.txt
flask>=3.0.0
jinja2>=3.1.0
playwright>=1.40.0
pyyaml>=6.0
pypdf>=3.0.0
pillow>=10.0.0
